+++ Source/Models/ResponseModels/Resource.cs
@@ -38,6 +38,8 @@ namespace BingMapsRESTToolkit
     [KnownType(typeof(SeaLevelData))]
     [KnownType(typeof(CompressedPointList))]
     [KnownType(typeof(GeospatialEndpointResponse))]
+    [KnownType(typeof(DistanceMatrix))]
+    [KnownType(typeof(DistanceMatrixAsyncStatus))]
     public class Resource
     {
         /// <summary>